backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9

# PDF parsing
pymupdf==1.24.10
pypdf==3.17.4

# Embeddings / ANN
sentence-transformers==3.0.1
numpy==1.26.4
faiss-cpu==1.8.0.post1

# LLM + HTTP
openai>=1.40.0
httpx==0.27.2
tiktoken==0.7.0

# Translation + TTS
deep-translator==1.11.4
gTTS==2.5.1

# NEW: robust topic matching
rapidfuzz==3.9.6
